requirements.txt
Flask>=2.3.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-groq>=0.2.0
langchain-core>=0.2.0
tavily-python>=0.3.0
reportlab>=4.0.0
requests>=2.31.0
gunicorn>=20.1.0
